refactor(RacingCarGameWinnerSelector) : 이름 변경
- RacingCarGameWinnerSelector : RacingCarGame 접두사로 변경함으로써, 목적인 자동차 경주 게임의 우승자를 선정한다는 목적을 명확하게 한다.
- 또한 Generator에서 Selector로 변경하여 우승자를 생산하는 것이 아닌 선정한다는 의미를 명확하게 한다.
