+++ src/main/java/racingcar/factory/BasicRacingCarGameFactory.java
@@ -4,7 +4,7 @@ package racingcar.factory;
 import racingcar.service.RacingCarGame;
 import racingcar.util.generator.BasicRacingCarGenerator;
 import racingcar.util.generator.BasicRacingCarGameRandomGenerator;
-import racingcar.util.generator.BasicRacingCarWinnerGenerator;
+import racingcar.util.selector.BasicRacingCarGameWinnerSelector;
 import racingcar.util.parser.BasicRacingCarNameParser;
 import racingcar.util.transporter.BasicRacingCarTransporter;
 import racingcar.util.validator.BasicRacingCarGameValidator;
@@ -16,7 +16,7 @@ public class BasicRacingCarGameFactory {
                 new BasicRacingCarGameValidator(),
                 new BasicRacingCarGenerator(),
                 new BasicRacingCarTransporter(),
-                new BasicRacingCarWinnerGenerator(),
+                new BasicRacingCarGameWinnerSelector(),
                 new BasicRacingCarGameRandomGenerator()
         );
     }
